Separated tests for Luv/Lab.
